Several rendering improvements
improved ingame menu scaling
improved rect handling
improved amy handling

+++ objects/base/spinbox.cpp
@@ -8,7 +8,7 @@
 
 #include "3rd_party/oxygine-framework/oxygine/actor/ClipRectActor.h"
 
-SpinBox::SpinBox(qint32 width, qint32 min, qint32 max, Mode mode)
+SpinBox::SpinBox(qint32 width, qreal min, qreal max, Mode mode)
     : m_MinValue(min),
       m_MaxValue(max),
       m_Mode(mode)
